% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{el_categories}
\alias{el_categories}
\title{Extract categories from metadata files in a directory}
\usage{
el_categories(data_dir)
}
\arguments{
\item{data_dir}{Directory containing the metadata folders.}
}
\value{
A data.frame with ID_fiche and categories.
}
\description{
Extract categories from metadata files in a directory
}
