% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{el_data}
\alias{el_data}
\title{Extract all metadata from a directory of flashcards}
\usage{
el_data(data_dir)
}
\arguments{
\item{data_dir}{Directory containing the flashcard folders.}
}
\value{
A data.frame with all extracted metadata.
}
\description{
Extract all metadata from a directory of flashcards
}
