% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tables.R
\name{el_data_summary}
\alias{el_data_summary}
\title{Summarizes metadata with counts of authors, organizations, keywords}
\usage{
el_data_summary(data)
}
\arguments{
\item{data}{the metadata, obtained via el_data()}
}
\value{
A data.frame with one row per flashcard and counts of authors, organizations, keywords
}
\description{
Summarizes metadata with counts of authors, organizations, keywords
}
