% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time.R
\name{el_durationhist}
\alias{el_durationhist}
\title{Create a histogram of the duration of flashcard-related projects}
\usage{
el_durationhist(table_time)
}
\arguments{
\item{data}{the table returned by el_table_time()}
}
\description{
Create a histogram of the duration of flashcard-related projects
}
