% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tables.R
\name{el_freq_keywords}
\alias{el_freq_keywords}
\title{Collects keywords table with their frequency}
\usage{
el_freq_keywords(data)
}
\arguments{
\item{data}{the metadata, obtained via el_data()}
}
\value{
A data.frame with keywords and their frequency
}
\description{
Collects keywords table with their frequency
}
