% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{el_get_auteurs}
\alias{el_get_auteurs}
\title{Extracts authors from a metadata XML object}
\usage{
el_get_auteurs(myxml)
}
\arguments{
\item{myxml}{An XML object.}
}
\value{
A data.frame with ID_fiche, R_personnes, and R_orgs.
}
\description{
Extracts authors from a metadata XML object
}
