% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{el_get_categories}
\alias{el_get_categories}
\title{Extraire les catégories d'un ensemble de fiches metadata#'
Lit les fichiers info.xml et récupère les catégories associées à chaque fiche.}
\usage{
el_get_categories(metadata_dir)
}
\arguments{
\item{metadata_dir}{Dossier où se trouvent les sous-dossiers contenant info.xml.}
}
\value{
Un data.frame avec les colonnes ID_fiche et categories.
}
\description{
Extraire les catégories d'un ensemble de fiches metadata#'
Lit les fichiers info.xml et récupère les catégories associées à chaque fiche.
}
