% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{el_get_info}
\alias{el_get_info}
\title{Extract metadata information from an XML object}
\usage{
el_get_info(myxml)
}
\arguments{
\item{myxml}{An XML object.}
}
\value{
A data.frame containing ID_fiche, debut, fin, xmin, xmax, ymin, ymax, supp, nlinks, ntables.
}
\description{
Extract metadata information from an XML object
}
