% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{el_get_keywords}
\alias{el_get_keywords}
\title{Extracts keywords from a metadata XML object}
\usage{
el_get_keywords(myxml)
}
\arguments{
\item{myxml}{An XML object.}
}
\value{
A data.frame with ID_fiche, keywords, and type.
}
\description{
Extracts keywords from a metadata XML object
}
