% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{el_get_metadata}
\alias{el_get_metadata}
\title{Extraire toutes les métadonnées d'un répertoire
Parcourt toutes les fiches dans un dossier et assemble les informations, auteurs et mots clés.}
\usage{
el_get_metadata(metadata_dir)
}
\arguments{
\item{metadata_dir}{Chemin du dossier contenant les fiches metadata.}
}
\value{
Un data.frame complet avec toutes les métadonnées fusionnées.
}
\description{
Extraire toutes les métadonnées d'un répertoire
Parcourt toutes les fiches dans un dossier et assemble les informations, auteurs et mots clés.
}
