% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{el_get_xml}
\alias{el_get_xml}
\title{Charger un fichier metadata.xml
Charge le fichier XML principal d'une fiche metadata dans son dossier.}
\usage{
el_get_xml(xmlpath)
}
\arguments{
\item{xmlpath}{Chemin du dossier de la fiche metadata.}
}
\value{
Un objet XML.
}
\description{
Charger un fichier metadata.xml
Charge le fichier XML principal d'une fiche metadata dans son dossier.
}
