% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{el_graph}
\alias{el_graph}
\title{Plot graph}
\usage{
el_graph(graph_data, shorten_name = FALSE)
}
\arguments{
\item{graph_data}{the graph data as obtained via el_graph_data()}

\item{shorten_name}{logical, whether to shorten the names of nodes (default: FALSE)}
}
\value{
the graph plot
}
\description{
Plot graph
}
