% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{el_graph_data}
\alias{el_graph_data}
\title{Build graph data}
\usage{
el_graph_data(data, nodetype)
}
\arguments{
\item{data}{the dataset as obtained via el_data()}

\item{nodetype}{either "personne" or "organisation", depending on whether you want to build a graph of people or organizations}
}
\value{
the graph data
}
\description{
Build graph data
}
