% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{el_map}
\alias{el_map}
\title{Génération d'une carte de densité de recouvrement des bounding box}
\usage{
el_map(map_data, mode = "plot", fond = fond_france)
}
\arguments{
\item{map_data}{Objet sf contenant une grille avec la densité de recouvrement des bounding box}

\item{mode}{Mode de visualisation: "plot" pour une carte statique avec tmap (valeur par défaut), "view" pour une carte interactive avec leaflet.}
}
\description{
Génération d'une carte de densité de recouvrement des bounding box
}
