% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{el_map_data}
\alias{el_map_data}
\title{Génération d'une grille de densité de recouvrement des bounding box}
\usage{
el_map_data(metadata)
}
\arguments{
\item{metadata}{Dataframe de métadonnées contenant les colonnes xmin, xmax, ymin, ymax}
}
\value{
Objet sf contenant une grille avec la densité de recouvrement des bounding box
}
\description{
Génération d'une grille de densité de recouvrement des bounding box
}
