% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tables.R
\name{el_metadata_summary}
\alias{el_metadata_summary}
\title{Fournir un tableau avec une ligne=une fiche}
\usage{
el_metadata_summary(metadata)
}
\arguments{
\item{metadata}{les métadonnées, obtenues via el_get_metadata()}
}
\value{
Un data.frame avec un comptage des auteurs, organisations, mots-clés
}
\description{
Fournir un tableau avec une ligne=une fiche
}
