% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/el_plot_categories.R
\name{el_plot_categories}
\alias{el_plot_categories}
\title{Plots frequencies of categories}
\usage{
el_plot_categories(categories, color = "#04a3bb")
}
\arguments{
\item{categories}{the data.frame with categories, obtained via el_categories()}
}
\value{
a plot with counts of categories
}
\description{
Plots frequencies of categories
}
