% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tables.R
\name{el_table_keywords}
\alias{el_table_keywords}
\title{Fournir les comptages de mots-clés}
\usage{
el_table_keywords(metadata)
}
\arguments{
\item{metadata}{les métadonnées, obtenues via el_get_metadata()}
}
\value{
Un data.frame avec les mots-clés, leur type, leur fréquence
}
\description{
Fournir les comptages de mots-clés
}
