% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time.R
\name{el_table_time}
\alias{el_table_time}
\title{Create a table with time information for each flashcard}
\usage{
el_table_time(metadata)
}
\arguments{
\item{metadata}{A data frame containing metadata information}
}
\description{
Create a table with time information for each flashcard
}
