% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time.R
\name{el_time_duration}
\alias{el_time_duration}
\title{Create a histogram of the duration of flashcard-related projects}
\usage{
el_time_duration(time_table, color = "#04a3bb")
}
\arguments{
\item{color}{the filling color. Defaults to "#04a3bb"}

\item{data}{the table returned by el_time_table()}
}
\description{
Create a histogram of the duration of flashcard-related projects
}
