% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time.R
\name{el_time_hist}
\alias{el_time_hist}
\title{Create a histogram of the number of flashcards per year}
\usage{
el_time_hist(time_table, color = "#04a3bb")
}
\arguments{
\item{time_table}{the table returned by el_time_table()}

\item{color}{the filling color. Defaults to "#04a3bb"}
}
\description{
Create a histogram of the number of flashcards per year
}
