% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time.R
\name{el_time_line}
\alias{el_time_line}
\title{Create a time_line plot}
\usage{
el_time_line(time_table, color = "#04a3bb")
}
\arguments{
\item{time_table}{the table returned by el_time_table()}

\item{color}{the color. Defaults to "#04a3bb"}
}
\description{
Create a time_line plot
}
