% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time.R
\name{el_time_table}
\alias{el_time_table}
\title{Create a table with time information for each flashcard}
\usage{
el_time_table(data)
}
\arguments{
\item{data}{A data frame containing metadata information}
}
\description{
Create a table with time information for each flashcard
}
