% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time.R
\name{el_timehist}
\alias{el_timehist}
\title{Create a histogram of the number of flashcards per year}
\usage{
el_timehist(table_time)
}
\arguments{
\item{data}{the table returned by el_table_time()}
}
\description{
Create a histogram of the number of flashcards per year
}
