% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordcloud.R
\name{el_wordcloud}
\alias{el_wordcloud}
\title{Nuage des mots}
\usage{
el_wordcloud(
  table_keywords,
  keywords_to_remove,
  max_size = 10,
  low_color = "#442484",
  high_color = "#04a3bb"
)
}
\arguments{
\item{table_keywords}{table_keywords, obtenues via el_table_keywords()}

\item{keywords_to_remove}{a vector of keywords to remove from the cloud}

\item{max_size}{the maximum size of a word}

\item{low_color}{the color of most frequent word (defaults to  "#442484")}

\item{high_color}{the color of least frequent word (defaults to "#04a3bb")}
}
\value{
Un data.frame avec les mots-clés, leur type, leur fréquence
}
\description{
Nuage des mots
}
\examples{
metadata_keywords \%>\%
dplyr::filter(type=="FREE")
el_wordcloud(metadata_keywords_FREE,keywords_to_remove=c("OHM VR","Rhône"))
}
