% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{fbox}
\alias{fbox}
\title{Création d'une bounding box en objet sf}
\usage{
fbox(nelat, nelng, swlat, swlng)
}
\arguments{
\item{nelat}{Latitude nord-est}

\item{nelng}{Longitude nord-est}

\item{swlat}{Latitude sud-ouest}

\item{swlng}{Longitude sud-ouest}
}
\description{
Création d'une bounding box en objet sf
}
