% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{fdate}
\alias{fdate}
\title{Extracts date from formatted ISO date strings}
\usage{
fdate(chain)
}
\arguments{
\item{chain}{A vector ISO formatted date strings.}
}
\value{
A vector of dates in "YYYY-MM-DD" format.
}
\description{
Extracts date from formatted ISO date strings
}
