% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{NAise}
\alias{NAise}
\title{Convertir un vecteur vide en NA
Renvoie NA si le vecteur est de longueur zéro, sinon renvoie le vecteur tel quel.}
\usage{
NAise(x)
}
\arguments{
\item{x}{Un vecteur quelconque.}
}
\value{
NA ou le vecteur d'origine.
}
\description{
Convertir un vecteur vide en NA
Renvoie NA si le vecteur est de longueur zéro, sinon renvoie le vecteur tel quel.
}
