% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{el_get_auteurs}
\alias{el_get_auteurs}
\title{Extraire les auteurs d'une fiche metadata}
\usage{
el_get_auteurs(myxml)
}
\arguments{
\item{myxml}{Un objet XML.}
}
\value{
Un data.frame contenant ID_fiche, R_personnes et R_orgs.
}
\description{
Récupère les noms des personnes et des organisations associées à la ressource.
}
