% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{el_get_keywords}
\alias{el_get_keywords}
\title{Extraire les mots clés d'une fiche metadata
Récupère les mots clés, leur type (FREE, INSPIRE ou ISO) et l'identifiant associé.}
\usage{
el_get_keywords(myxml)
}
\arguments{
\item{myxml}{Un objet XML.}
}
\value{
Un data.frame avec ID_fiche, keywords et types.
}
\description{
Extraire les mots clés d'une fiche metadata
Récupère les mots clés, leur type (FREE, INSPIRE ou ISO) et l'identifiant associé.
}
