% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{el_get_xml}
\alias{el_get_xml}
\title{Gets the XML object from a metadata folder}
\usage{
el_get_xml(xmlpath)
}
\arguments{
\item{xmlpath}{Path to the folder containing the metadata.}
}
\value{
An XML object read from the metadata.xml file.
}
\description{
Gets the XML object from a metadata folder
}
