% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{el_graph}
\alias{el_graph}
\title{Trace le graphe}
\usage{
el_graph(graph_data, shorten_name = FALSE)
}
\arguments{
\item{graph_data}{le jeu de données du graphe, tel qu'obtenu avec el_graph_data()}

\item{shorten_name}{booléen, si TRUE, raccourcit les noms des noeuds (utile pour les organisations avec des noms longs)}
}
\value{
le graphe
}
\description{
Trace le graphe
}
