% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{el_graph_data}
\alias{el_graph_data}
\title{Construis le jeu de données pour le graphe}
\usage{
el_graph_data(metadata, nodetype)
}
\arguments{
\item{metadata}{the metadata dataframe}

\item{nodetype}{la nature du noeud ("personne" ou "organisation")}
}
\value{
un tableau avec le nombre de liaisons (nombre de fiches) entre les futurs noeuds du graphe (organisations ou personnes)
}
\description{
Construis le jeu de données pour le graphe
}
