% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{el_map}
\alias{el_map}
\title{Generates a map visualizing the density of bounding box coverage}
\usage{
el_map(map_data, mode = "plot")
}
\arguments{
\item{map_data}{an sf object containing a grid with the density of bounding box coverage}

\item{mode}{"plot" for static map (tmap) or "view" for interactive map (leaflet)}
}
\description{
Generates a map visualizing the density of bounding box coverage
}
