% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{el_map_data}
\alias{el_map_data}
\title{Generates a grid with the density of bounding box coverage}
\usage{
el_map_data(data)
}
\arguments{
\item{data}{Dataframe containing metadata with bounding box coordinates}
}
\value{
an sf object containing a grid with the density of bounding box coverage
}
\description{
Generates a grid with the density of bounding box coverage
}
