% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordcloud.R
\name{el_wordcloud}
\alias{el_wordcloud}
\title{Create a wordcloud from a table of keywords}
\usage{
el_wordcloud(
  table_keywords,
  keywords_to_remove,
  max_size = 10,
  low_color = "#442484",
  high_color = "#04a3bb"
)
}
\arguments{
\item{table_keywords}{a data.frame with columns 'keywords' and 'freq'}

\item{keywords_to_remove}{a vector of keywords to remove from the wordcloud}

\item{max_size}{the maximum size of a word}

\item{low_color}{the color of most frequent word (defaults to  "#442484")}

\item{high_color}{the color of least frequent word (defaults to "#04a3bb")}
}
\value{
a wordcloud
}
\description{
Create a wordcloud from a table of keywords
}
\examples{
metadata_keywords \%>\%
dplyr::filter(type=="FREE")
el_wordcloud(metadata_keywords_FREE,keywords_to_remove=c("OHM VR","Rhône"))
}
