% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{fbox}
\alias{fbox}
\title{Creation of a sf bounding box}
\usage{
fbox(nelat, nelng, swlat, swlng)
}
\arguments{
\item{nelat}{north-east latitude}

\item{nelng}{north-east longitude}

\item{swlat}{south-west latitude}

\item{swlng}{south-west longitude}
}
\description{
Creation of a sf bounding box
}
