% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{fdate}
\alias{fdate}
\title{Extraire la date d'un vecteur de chaînes ISO
Sépare une chaîne de type "YYYY-MM-DDTHH:MM:SS" et renvoie uniquement la date.}
\usage{
fdate(chain)
}
\arguments{
\item{chain}{Un vecteur de chaînes contenant des dates ISO avec un T.}
}
\value{
Un vecteur de dates sous forme de chaînes "YYYY-MM-DD".
}
\description{
Extraire la date d'un vecteur de chaînes ISO
Sépare une chaîne de type "YYYY-MM-DDTHH:MM:SS" et renvoie uniquement la date.
}
