% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{NAise}
\alias{NAise}
\title{Converts a zero-length vector to NA}
\usage{
NAise(x)
}
\arguments{
\item{x}{A vector.}
}
\value{
NA if the vector has length 0, otherwise returns the vector.
}
\description{
Converts a zero-length vector to NA
}
